Correct for over-enthusiastic compiler warnings.
git-svn-id: http://code.icecube.wisc.edu/svn/projects/photospline/trunk@68135 16731396-06f5-0310-8873-f7f720988828

+++ private/util/evalsplinefits.c
@@ -35,7 +35,7 @@ int main(int argc, char **argv) {
     #ifdef TIMING
 	tp2.tv_usec += (tp2.tv_sec - tp1.tv_sec)*1e6;
 	printf("Time to open table: %ld.%06ld seconds\n", 
-	    tp2.tv_sec - tp1.tv_sec, tp2.tv_usec - tp1.tv_usec);
+	    (long)(tp2.tv_sec - tp1.tv_sec), tp2.tv_usec - tp1.tv_usec);
     #endif
 
 	if (argc < 2+table.ndim)
